+++ src/Messaging/MySqlMessageQueue.php
@@ -310,7 +310,7 @@ class MySqlMessageQueue implements MessageQueueInterface
 
         $table = $schema->createTable($this->table);
 
-        $table->addColumn('id', 'integer', [
+        $table->addColumn('id', 'bigint', [
             'autoincrement' => true,
             'unsigned'      => true,
             'notnull'       => true
